Update Addition.js
## Description

Updated the overall code for better code understandability.

Earlier the function add(a, b) directly returns the expression a + b and does not store the value of expression a + b in  another variable. 

Now, It stores the result of addition into another variable and then returns it.

Earlier we calling the addition function directly into the print statement. Now it first stores the answer and then print's it.